requirements.txt
youtube-transcript-api
streamlit
google-generativeai
python-dotenv
pathlib
bs4
